Renamed flowTest CLI command to protocolTest
